+++ src/modules/comms/dto/create-comm.dto.ts
@@ -85,6 +85,12 @@ export class CreateCommDto {
 
   @ApiPropertyOptional()
   @IsOptional()
+  @IsInt()
+  @Type(() => Number)
+  documentTypeId?: number;
+
+  @ApiPropertyOptional()
+  @IsOptional()
   @IsArray()
   @ValidateNested({ each: true })
   @Type(() => ActionDto)